Fix build of truss for CheriABI.
- Force address interpretation so I don't have to sprinkle __cheri_addr
  casts all over the place.
- Add a print_pointer helper function to print a raw pointer argument
  and use it everywhere that prints a syscall arg as a raw pointer.
  For now this just uses %p but eventually we can augment this function
  to print out a format akin to what GDB uses with bounds, etc.

+++ usr.bin/truss/setup.c
@@ -559,7 +559,7 @@ exit_syscall(struct trussinfo *info, struct ptrace_lwpinfo *pl)
 			 */
 			if (psr.sr_error != 0) {
 				asprintf(&temp, "0x%lx",
-				    t->cs.args[sc->args[i].offset]);
+				    (unsigned long)t->cs.args[sc->args[i].offset]);
 			} else {
 				temp = print_arg(&sc->args[i],
 				    t->cs.args, psr.sr_retval, info);
